+++ src/views_frames_summarize/__init__.py
@@ -17,6 +17,8 @@ from views_frames_summarize.aggregate import (
 )
 from views_frames_summarize.bimodality import bimodality
 from views_frames_summarize.collapse import collapse
+from views_frames_summarize.exceedance import exceedance, exceedance_reducer
+from views_frames_summarize.expected_shortfall import expected_shortfall
 from views_frames_summarize.interval import hdi, quantiles
 from views_frames_summarize.point import map_estimate
 from views_frames_summarize.summarize_tower import TowerSummary, summarize_tower
@@ -29,6 +31,9 @@ __all__ = [
     "aggregate_distributions_arrays",
     "bimodality",
     "collapse",
+    "exceedance",
+    "exceedance_reducer",
+    "expected_shortfall",
     "hdi",
     "hdi_tower",
     "map_estimate",